requirements.txt
qiskit-aqua>=0.6.0,<0.7.0
pyobjc-core; sys_platform == 'darwin'
pyobjc-framework-Cocoa; sys_platform == 'darwin'